# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LicitaIA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /processes:
    get:
      operationId: listProcesses
      tags: [processes]
      summary: List all processes
      responses:
        "200":
          description: List of processes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Process"
    post:
      operationId: createProcess
      tags: [processes]
      summary: Create a new process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessBody"
      responses:
        "201":
          description: Created process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /processes/{id}:
    get:
      operationId: getProcess
      tags: [processes]
      summary: Get a process by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Process details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProcess
      tags: [processes]
      summary: Update a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProcessBody"
      responses:
        "200":
          description: Updated process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProcess
      tags: [processes]
      summary: Delete a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /processes/{id}/requirements:
    get:
      operationId: listRequirements
      tags: [requirements]
      summary: List extracted requirements for a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of requirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExtractedRequirement"
  /processes/{id}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files uploaded for a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UploadedFile"
  /processes/{id}/validation:
    get:
      operationId: listValidationItems
      tags: [validation]
      summary: List validation items (checklist) for a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation checklist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationItem"
...
  /processes/{id}/report:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get the final report for a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Final report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Process:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        agency:
          type: string
        modality:
          type: string
        editalNumber:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum:
            - criado
            - edital_enviado
            - edital_processando
            - exigencias_extraidas
            - aguardando_documentos
            - documentos_enviados
            - em_conferencia
            - pendencias_encontradas
            - pronto_para_revisao
            - concluido
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - agency
        - modality
        - status
        - createdAt
        - updatedAt
    ProcessDetail:
      allOf:
        - $ref: "#/components/schemas/Process"
        - type: object
          properties:
            editalFile:
              oneOf:
                - $ref: "#/components/schemas/UploadedFile"
                - type: "null"
            documentFiles:
              type: array
              items:
                $ref: "#/components/schemas/UploadedFile"
            requirements:
              type: array
              items:
                $ref: "#/components/schemas/ExtractedRequirement"
            validationItems:
              type: array
              items:
                $ref: "#/components/schemas/ValidationItem"
            report:
              oneOf:
                - $ref: "#/components/schemas/FinalReport"
                - type: "null"
          required:
            - documentFiles
            - requirements
            - validationItems
    CreateProcessBody:
      type: object
      properties:
        title:
          type: string
        agency:
          type: string
        modality:
          type: string
        editalNumber:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - title
        - agency
        - modality
    UpdateProcessBody:
      type: object
      properties:
        title:
          type: string
        agency:
          type: string
        modality:
          type: string
        editalNumber:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum:
            - criado
            - edital_enviado
            - edital_processando
            - exigencias_extraidas
            - aguardando_documentos
            - documentos_enviados
            - em_conferencia
            - pendencias_encontradas
            - pronto_para_revisao
            - concluido
    UploadedFile:
      type: object
      properties:
        id:
          type: integer
        processId:
          type: integer
        fileType:
          type: string
          enum: [edital, anexo, documento_empresa]
        name:
          type: string
        path:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        uploadedAt:
          type: string
      required:
        - id
        - processId
        - fileType
        - name
        - path
        - mimeType
        - size
        - uploadedAt
    ExtractedRequirement:
      type: object
      properties:
        id:
          type: integer
        processId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        mandatory:
          type: boolean
        sourceExcerpt:
          type: ["string", "null"]
        sourcePage:
          type: ["integer", "null"]
        confidence:
          type: number
        needsReview:
          type: boolean
      required:
        - id
        - processId
        - title
        - mandatory
        - confidence
        - needsReview
    SubmittedDocument:
      type: object
      properties:
        id:
          type: integer
        processId:
          type: integer
        fileId:
          type: integer
        detectedType:
          type: ["string", "null"]
        detectedDate:
          type: ["string", "null"]
        detectedValidity:
          type: ["string", "null"]
        confidence:
          type: number
        needsReview:
          type: boolean
      required:
        - id
        - processId
        - fileId
        - confidence
        - needsReview
    ValidationItem:
      type: object
      properties:
        id:
          type: integer
        processId:
          type: integer
        requirementId:
          type: integer
        submittedDocumentId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [ok, faltando, vencido, divergente, revisar]
        notes:
          type: ["string", "null"]
        requirement:
          $ref: "#/components/schemas/ExtractedRequirement"
        submittedDocument:
          oneOf:
            - $ref: "#/components/schemas/SubmittedDocument"
            - type: "null"
      required:
        - id
        - processId
        - requirementId
        - status
        - requirement
...
    FinalReport:
      type: object
      properties:
        id:
          type: integer
        processId:
          type: integer
        summary:
          type: string
        okCount:
          type: integer
        missingCount:
          type: integer
        expiredCount:
          type: integer
        divergentCount:
          type: integer
        reviewCount:
          type: integer
        nextSteps:
          type: ["string", "null"]
        generatedAt:
          type: string
      required:
        - id
        - processId
        - summary
        - okCount
        - missingCount
        - expiredCount
        - divergentCount
        - reviewCount
        - generatedAt
...