# lib/api-zod/src/generated/types/process.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LicitaIA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessStatus } from "./processStatus";

export interface Process {
  id: number;
  title: string;
  agency: string;
  modality: string;
  /** @nullable */
  editalNumber?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  notes?: string | null;
  status: ProcessStatus;
  createdAt: string;
  updatedAt: string;
}