# lib/api-zod/src/generated/types/processDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LicitaIA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedRequirement } from "./extractedRequirement";
import type { FinalReport } from "./finalReport";
import type { Process } from "./process";
import type { UploadedFile } from "./uploadedFile";
import type { ValidationItem } from "./validationItem";

export type ProcessDetail = Process & {
  editalFile?: UploadedFile | null;
  documentFiles: UploadedFile[];
  requirements: ExtractedRequirement[];
  validationItems: ValidationItem[];
  report?: FinalReport | null;
};