# lib/db/src/schema/processes.ts
import { pgTable, text, serial, timestamp, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const processStatusEnum = [
  "criado",
  "edital_enviado",
  "edital_processando",
  "exigencias_extraidas",
  "aguardando_documentos",
  "documentos_enviados",
  "em_conferencia",
  "pendencias_encontradas",
  "pronto_para_revisao",
  "concluido",
] as const;

export const processesTable = pgTable("processes", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  agency: text("agency").notNull(),
  modality: text("modality").notNull(),
  editalNumber: text("edital_number"),
  deadline: text("deadline"),
  notes: text("notes"),
  status: text("status").notNull().default("criado"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertProcessSchema = createInsertSchema(processesTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
});

export type InsertProcess = z.infer<typeof insertProcessSchema>;
export type Process = typeof processesTable.$inferSelect;